Create V1Beta3 class
